Fix compilation errors in r26328
git-svn-id: http://root.cern.ch/svn/root/trunk@26329 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint7/src/disp.cxx
@@ -732,7 +732,7 @@ static int G__display_classinheritance(FILE *fout, int tagnum, char *space)
 
    for (i = 0;i < baseclass->basen;i++) {
       if (baseclass->property[i]&G__ISDIRECTINHERIT) {
-         sprintf(msg, "%s0x%-8lx ", space , baseclass->baseoffset[i]);
+         sprintf(msg, "%s0x%-8lx ", space , (size_t)baseclass->baseoffset[i]);
          if (G__more(fout, msg)) return(1);
          if (baseclass->property[i]&G__ISVIRTUALBASE) {
             sprintf(msg, "virtual ");
@@ -1492,7 +1492,7 @@ int Cint::Internal::G__objectmonitor(FILE *fout, char *pobject, const ::Reflex::
 #endif
          }
          else {
-            sprintf(msg, "%s0x%-8lx ", space , baseclass->baseoffset[i]);
+            sprintf(msg, "%s0x%-8lx ", space , (size_t)baseclass->baseoffset[i]);
             if (G__more(fout, msg)) return(1);
             msg[0] = 0;
             switch (baseclass->baseaccess[i]) {
@@ -1601,7 +1601,7 @@ int Cint::Internal::G__varmonitor(FILE* fout, const ::Reflex::Scope scope, const
       //
       //  Print out member address.
       //
-      sprintf(msg, "0x%-8lx ", addr);
+      sprintf(msg, "0x%-8lx ", (size_t)addr);
       if (G__more(fout, msg)) {
          return 1;
       }
@@ -1925,10 +1925,10 @@ int Cint::Internal::G__varmonitor(FILE* fout, const ::Reflex::Scope scope, const
                   //  Print the member address and the char string.
                   //
                   if (isprint(*(char*)addr)) {
-                     sprintf(msg, "=0x%lx=\"%s\"", addr, (char*) addr); // FIXME: This can fail, a char array may not be zero terminated!  And we do not check the whole array for printability.
+                     sprintf(msg, "=0x%lx=\"%s\"", (size_t)addr, (char*) addr); // FIXME: This can fail, a char array may not be zero terminated!  And we do not check the whole array for printability.
                   }
                   else {
-                     sprintf(msg, "=0x%lx", addr);
+                     sprintf(msg, "=0x%lx", (size_t)addr);
                   }
                   if (G__more(fout, msg)) {
                      return 1;
@@ -1939,7 +1939,7 @@ int Cint::Internal::G__varmonitor(FILE* fout, const ::Reflex::Scope scope, const
                   //
                   //  Print the member address only.
                   //
-                  sprintf(msg, "=0x%lx", addr);
+                  sprintf(msg, "=0x%lx", (size_t)addr);
                   if (G__more(fout, msg)) {
                      return 1;
                   }
